+++ core/src/test/scala/eu/monniot/scala3mock/mock/MockSuite.scala
@@ -1,7 +1,7 @@
 package mock
 
 import eu.monniot.scala3mock.ScalaMocks
-import eu.monniot.scala3mock.main.Default
+import eu.monniot.scala3mock.Default
 import fixtures.*
 
 class MockSuite extends munit.FunSuite with ScalaMocks {
